refactored the files to conform with conventions
